third_party/WebKit/LayoutTests/platform/win/external/csswg-test/css-flexbox-1/getcomputedstyle/flexbox_computedstyle_align-self-invalid-expected.txt

This is a testharness.js-based test.
FAIL flexbox | computed style | align-self: invalid assert_equals: expected "stretch" but got "normal"
Harness: the test ran to completion.
